add my.cnf
Add config file to make MySQL play nice with OSX. MySQL 5.6 changed the default
behavior to use a file per table, which has the potential to exceed OSX's limit
on the number of files a process can have open (this crops up mostly when using
parallel_specs). Fix this by telling MySQL NOT to do the file per table thing.