mysql default timezone=UTC
Set MySQL's default timezone to UTC. Since we're working under the assumption
that times are stored in UTC, make sure MySQL works with the same understanding.
This is critical if comparing a time that was persisted as UTC to a calculated
value like NOW(), which uses the currently set time zone.