use slash as right shift modifer
The right Shift key requires an extra keywidth of pinky stretch compared to the left Shift
key. Try using `/` as a supplimentary Shift key, so it acts as Shift when pressed in
concert with another key, and continues to send `/` when pressed alone.

Karabiner provides an empty file in which you can define custom tweaks like this in:
  ~/Library/Application\ Support/private.xml

I'm moving this file into my dotfiles repo and symlinking to it.