introduce etags git hook
Use a similar strategy for generating etags as I was for ctags. I'm excluding json (which
universal-ctags added support for) because it results in etags of monstrous proportions.
Might as well exclude it from ctags as well, as I don't see myself using it often.

Once I feel comfortable with my etags configuration, I'll add etags to the git lifecycle
hooks.