update ctags & etags git hooks
The universal-ctags project has added tagging support for more languages, but I don't want
to generate tags for all of them, svg for example. The etags implementation currently has
a bug in that it doesn't truncate source line references. This is problematic for long
lines, such as would occur in a minified file. Thus, I'm excluding css from my etags until
the bug is fixed.

The change from -f to -o is a no-op, they're aliases, I just prefer -o.