improve bin/e Emacs cli-starter
This variation is callable with or without file args, and will make emacs visible even if
it was previously hidden.

The `if [ $# -gt 0]` block that handles file args is run in the background (see the
closing `fi &`), but its contents run synchronously with respect to each other. That
prevents deleting the tempfiles before they've been visited, without resorting to `sleep`.