Less, lesskey: fix ctrl-arrows and shift-arrows
To make Less recognize {Ctrl,Shift}-{right,left} keys, put the appropriate escape
sequences in a lesskey file and run the lesskey program to compile it into a binary config
file that Less will look for. Also, needed to brew install Less, since mac comes with one
that's been built without support for using a key file. Then symlink ~/.less to the less
file in this repo, and set a MANPAGER env var to use the brewed version of Less (otherwise
running `less` will use the brewed version, but running `man` will not). I might end up
setting the overall PAGER env var instead, but thought I'd start with this more specific
setting and see if that's sufficient. Altogether that's:

  brew install less
  vi ~/.lesskey
  lesskey
  mv ~/.lesskey ~/Development/resources/dotfiles/lesskey
  mv ~/.less ~/Development/resources/dotfiles/less
  ln -svhF ~/Development/resources/dotfiles/less ~/.less