bashrc, bash_profile
I'm starting to gain a better understanding of what should go in
bash_profile and what in bashrc. For one, environment variables should
probably be set in bash_profile, since that's responsible for
configuring the overall environment, whereas bashrc is intended more
specifically for initializing an interactive shell session. Although
the default behavior under macOS is to run every interactive shell as a
login shell, the general idea is to set up the overall environment in
bash_profile and avoid doing so all over again in each interactive
session you spawn.