set skip-completed-text on
This is so cool! One of my favorite Emacs features (that I wish Vim had
as well): completing from the middle of a word works correctly, it moves
over the existing characters to the right of the cursor, then finishes
the rest of the word. That, as opposed to inserting the rest of the word
from the current cursor position and ending up with extra cruft at the
end of the word.