change `jobs` alias: b -> jb
Somehow I find this *easier* to type. I think it's because the `b` is in
an awkward position, yet my finger knows how to find it when already in
the midst of typing. So I can hit it consistently in the middle/end of a
word, but tend to choke/miss it if using it alone or at the start of a
word.