bash_history_filter.awk: condense the braces
This just an aesthetic change, but compressing the syntax a bit made
something jump out more clearly: before we've reached the first
timestamped line in the file, we allow every line in unfiltered,
regardless of the exclusion pattern. I'm actually ok with that. It
simplifies the code significantly at the cost of overlooking exclusions
in non-timestamped sections of history, but since I've turned on
timestamps, it's only a matter of time until that section is pushed out
of the file by incoming lines of history.