PROMPT_COMMAND='maybe_newline'
Pull the conditional newline logic out of PS1 and into a PROMPT_COMMAND.
This still suffers from broken typeahead, because in order to determine
the cursor position, cursor_col() sends a terminal control sequence to
the input stream. Any input that was typed prior to that gets lost.

Without the call to cursor_col(), I could run a command, and while it's
still printing output, I could manually type `ls`. When the original
command's output finishes printing, Bash will print a fresh prompt and
put the `ls` on the command line.

With the call to cursor_col(), I'd run the original command, type an
`l` while the command output is still printing to the terminal, the
printing finishes, cursor_col() runs, I type an `s` before Bash has
printed a fresh prompt, then Bash prints the prompt and puts the `s`
on the command line. The `l` that preceded it, though, has been lost,
since it occured before cursor_col() sent the control sequence to the
terminal.

Nonetheless, this change at least separates the newline logic from the
prompt content.