inputrc: nop some modified keys
Instead of binding these to their unmodified variants, make them nops.
This fixes the behavior when using such a key as an alacritty
keybinding. For example, binding Ctrl-Shift-Space to alacritty's
ToggleViMode action was causing a space character to be inserted on the
command-line as well as toggling Vi mode. I still want to send the
escape sequences when these key combos are pressed, as that allows me to
use them in vim key maps, but in Bash I don't want them doing anything
other than what's been configured in the alacritty.yml keybindings.